Исправить баги

+++ ewm-service/src/main/java/ru/practicum/ewmservice/admin/service/AdminEventService.java
@@ -6,8 +6,8 @@ import ru.practicum.ewmservice.event.DTO.EventFullDTO;
 import java.util.List;
 
 public interface AdminEventService {
-    public List<EventFullDTO> getEvents(List<Long> ids, List<String> states, List<Integer> categories,
-                                        String rangeStart,String rangeEnd,int from, int size);
+    public List<EventFullDTO> getEvents(List<Integer> ids, List<String> states, List<Integer> categories,
+                                        String rangeStart, String rangeEnd, Integer from, Integer size);
 
     public EventFullDTO updateEvent(Integer eventId, UpdateEventAdminRequest updateEventDTO);
